feat: add fhir parser module
